tools/packager/requirements.txt
PyYAML>=5.4
psutil==5.8.0
pydantic==1.8.2
